corrected survey

+++ src/main/java/com/softbistro/survey/creating/survey/controller/SurveyController.java
@@ -151,11 +151,11 @@ public class SurveyController {
 	@RequestMapping(value = "/{survey_id}", method = RequestMethod.DELETE, produces = "application/json")
 	public ResponseEntity<Object> deleteSurvey(@PathVariable(value = "survey_id") Integer surveyId,
 			@RequestHeader String token) {
-		//
-		// if (!authorizationService.checkAccess(token)) {
-		//
-		// return new ResponseEntity<Object>(HttpStatus.UNAUTHORIZED);
-		// }
+
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<Object>(HttpStatus.UNAUTHORIZED);
+		}
 
 		return surveyService.deleteSurvey(surveyId);
 	}
@@ -167,7 +167,12 @@ public class SurveyController {
 	 * @return
 	 */
 	@RequestMapping(value = "/{survey_id}/start", method = RequestMethod.PUT)
-	public ResponseEntity<Object> startSurvey(@PathVariable(value = "survey_id") Integer surveyId) {
+	public ResponseEntity<Object> startSurvey(@PathVariable(value = "survey_id") Integer surveyId,
+			@RequestHeader String token) {
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<Object>(HttpStatus.UNAUTHORIZED);
+		}
 		return surveyService.startSurvey(surveyId);
 	}
 
@@ -178,7 +183,12 @@ public class SurveyController {
 	 * @return
 	 */
 	@RequestMapping(value = "/{survey_id}/stop", method = RequestMethod.PUT)
-	public ResponseEntity<Object> stopSurvey(@PathVariable(value = "survey_id") Integer surveyId) {
+	public ResponseEntity<Object> stopSurvey(@PathVariable(value = "survey_id") Integer surveyId,
+			@RequestHeader String token) {
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<Object>(HttpStatus.UNAUTHORIZED);
+		}
 		return surveyService.stopSurvey(surveyId);
 	}
 }
